Introduced Real typedef (in support of #13 and #14)

+++ Source/Graphics/ShaderDataValueFormat.cpp
@@ -55,7 +55,7 @@ UniformValue ShaderDataValueFormat::parseValue(UniformType type, const DataValue
     case UniformType::Texture:
         return UniformValue(dataValue.asInt(), type);
     case UniformType::Float:
-        return UniformValue((float)dataValue.asDouble());
+        return UniformValue(dataValue.asDouble());
     case UniformType::Vector2:
         return UniformValue(dataValue.asVector2());
     case UniformType::Vector3:
